+++ examples/relationships-gql/go.mod
@@ -1,0 +1,80 @@
+module github.com/goliatone/go-crud/examples/relationships-gql
+
+go 1.24.0
+
+require (
+	github.com/99designs/gqlgen v0.17.84
+	github.com/gofiber/fiber/v2 v2.52.6
+	github.com/goliatone/go-crud v0.0.0-00010101000000-000000000000
+	github.com/goliatone/go-crud/gql v0.0.0-00010101000000-000000000000
+	github.com/goliatone/go-repository-bun v0.9.0
+	github.com/goliatone/go-router v0.27.0
+	github.com/google/uuid v1.6.0
+	github.com/mattn/go-sqlite3 v1.14.28
+	github.com/uptrace/bun v1.2.14
+	github.com/uptrace/bun/dialect/sqlitedialect v1.2.14
+	github.com/vektah/gqlparser/v2 v2.5.31
+)
+
+require (
+	dario.cat/mergo v1.0.1 // indirect
+	github.com/agnivade/levenshtein v1.2.1 // indirect
+	github.com/andybalholm/brotli v1.1.1 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/ettle/strcase v0.2.0 // indirect
+	github.com/fasthttp/websocket v1.5.8 // indirect
+	github.com/flosch/pongo2/v6 v6.0.0 // indirect
+	github.com/gertd/go-pluralize v0.2.1 // indirect
+	github.com/go-ozzo/ozzo-validation/v4 v4.3.0 // indirect
+	github.com/go-viper/mapstructure/v2 v2.4.0 // indirect
+	github.com/gobwas/glob v0.2.3 // indirect
+	github.com/goccy/go-yaml v1.18.0 // indirect
+	github.com/gofiber/contrib/websocket v1.3.4 // indirect
+	github.com/gofiber/template v1.8.3 // indirect
+	github.com/gofiber/template/django/v3 v3.1.13 // indirect
+	github.com/gofiber/utils v1.1.0 // indirect
+	github.com/goliatone/go-composite-fs v0.0.1 // indirect
+	github.com/goliatone/go-errors v0.9.0 // indirect
+	github.com/goliatone/go-template v0.3.1 // indirect
+	github.com/goodsign/monday v1.0.2 // indirect
+	github.com/gorilla/websocket v1.5.1 // indirect
+	github.com/hashicorp/golang-lru/v2 v2.0.7 // indirect
+	github.com/jinzhu/inflection v1.0.0 // indirect
+	github.com/julienschmidt/httprouter v1.3.0 // indirect
+	github.com/klauspost/compress v1.17.11 // indirect
+	github.com/lib/pq v1.10.9 // indirect
+	github.com/mattn/go-colorable v0.1.14 // indirect
+	github.com/mattn/go-isatty v0.0.20 // indirect
+	github.com/mattn/go-runewidth v0.0.16 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/puzpuzpuz/xsync/v3 v3.5.1 // indirect
+	github.com/rivo/uniseg v0.4.7 // indirect
+	github.com/savsgio/gotils v0.0.0-20240303185622-093b76447511 // indirect
+	github.com/sosodev/duration v1.3.1 // indirect
+	github.com/stretchr/objx v0.5.2 // indirect
+	github.com/stretchr/testify v1.11.1 // indirect
+	github.com/tmthrgd/go-hex v0.0.0-20190904060850-447a3041c3bc // indirect
+	github.com/urfave/cli/v3 v3.6.1 // indirect
+	github.com/valyala/bytebufferpool v1.0.0 // indirect
+	github.com/valyala/fasthttp v1.54.0 // indirect
+	github.com/valyala/tcplisten v1.0.0 // indirect
+	github.com/vmihailenco/msgpack/v5 v5.4.1 // indirect
+	github.com/vmihailenco/tagparser/v2 v2.0.0 // indirect
+	golang.org/x/mod v0.30.0 // indirect
+	golang.org/x/net v0.47.0 // indirect
+	golang.org/x/sync v0.18.0 // indirect
+	golang.org/x/sys v0.38.0 // indirect
+	golang.org/x/text v0.31.0 // indirect
+	golang.org/x/time v0.8.0 // indirect
+	golang.org/x/tools v0.39.0 // indirect
+	gopkg.in/yaml.v2 v2.4.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
+
+replace github.com/goliatone/go-crud => ../..
+
+replace github.com/goliatone/go-crud/gql => ../../gql
+
+replace github.com/PuerkitoBio/goquery => github.com/PuerkitoBio/goquery v1.9.3
+
+replace github.com/andybalholm/cascadia => github.com/andybalholm/cascadia v1.3.2
